Added support for cases when package.* are directories
For certain cases it is beneficial to use directories instead of files
for

* package.accept_keywords
* package.license
* package.mask
* package.unmask

For example, it is a requirement for crossdev.

This commit adds a task in `tasks/check_if_dir.yml` that takes in a
variable `target_path`, checks if it is a directory.

If the target_path is a directory, the value of a new parameter
"portage_ansible_managed_filename" is appended to it, so all edits in
tasks are done to the file residing in

`{{ target_path }} / {{ portage_ansible_managed_filename }}`

The new parameter default value is "ansible_managed".

Tasks in:
* tasks/accept_keyword.yml
* tasks/accept_license.yml
* tasks/mask.yml
* tasks/unmask.yml

have been adjusted to use the new task.